src/ingestion/requirements.txt
pandas==2.3.0
python_box==7.3.2
PyYAML==6.0.2
